feat: remove unused kv store, cleanup chunk store
